chore(release): bump mcp-server 0.7.10
Publishes SMI-6109/SMI-6111's removal of SUPABASE_SERVICE_ROLE_KEY from the
customer-facing MCP server so the fix actually reaches design partners.

Also splits index.ts's runStartupDiagnostics() into index.startup-helpers.ts
(same pattern as SMI-5639's existing extraction) — the version-string bump
tripped the 500-LOC gate on a file that was already at the ceiling.

Refs SMI-6111
